SA-817: Move FilterSortCollection to shared component
