+++ src/components/collection/FilterSortCollection.module.scss
@@ -1,4 +1,4 @@
-@import '~@sparkpost/matchbox/src/styles/config.scss';
+@import '../../../node_modules/@sparkpost/matchbox/src/styles/config';
 
 .Title {
   padding: spacing(base) 1rem;
