+++ src/pages/blacklist/components/MonitorsCollection.js
@@ -3,7 +3,7 @@ import { Button } from '@sparkpost/matchbox';
 
 import { PageLink } from 'src/components';
 import styles from './MonitorsCollection.module.scss';
-import FilterSortCollection from 'src/pages/inboxPlacement/components/FilterSortCollection';
+import FilterSortCollection from 'src/components/collection/FilterSortCollection';
 
 const filterBoxConfig = {
   show: true,
